He cambiado long numero por int numero.

+++ src/main/java/es/hotel_back_v2/hotelV2/controllers/HabitacionController.java
@@ -41,7 +41,7 @@ public class HabitacionController {
     }
 
     @DeleteMapping("/eliminarHabitacion/{numero}")
-    public void eliminarHabitacion(@PathVariable Long numero) {
+    public void eliminarHabitacion(@PathVariable int numero) {
         habitacionService.eliminarHabitacion(numero);
     }
