+++ src/main/java/es/hotel_back_v2/hotelV2/services/HabitacionService.java
@@ -28,7 +28,7 @@ public class HabitacionService {
         return habitacionRepository.save(habitacion);
     }
 
-    public void eliminarHabitacion(Long numero) {
+    public void eliminarHabitacion(int numero) {
         habitacionRepository.deleteById(numero);
     }
